Change parameters

+++ Conversor/Conversor.java
@@ -1,6 +1,6 @@
 
 public interface Conversor {
 	
-	abstract double Convertir(double pUnidadInicial,double pUnidadFinal, double pDigito);
+	abstract double Convertir(int pUnidadInicial,int pUnidadFinal,int pDigito);
 	
 }
