+++ Conversor/ConversorDivisas.java
@@ -6,7 +6,7 @@ public class ConversorDivisas implements Conversor{
 	private double vDigito;
 	
 	@Override
-	public double Convertir(double pUnidadInicial, double pUnidadFinal, double pDigito) {
+	public double Convertir(int pUnidadInicial, int pUnidadFinal, double pDigito) {
 		return 0;
 	}
 }
